requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.4
django-crispy-forms==2.0
django-environ==0.10.0
django-simple-history==3.3.0
django-template-partials==23.2
six==1.16.0
sqlparse==0.4.4
